[#27272] Create and tested SwapController,Service and Provider

+++ src/main/java/com/cryptonita/app/data/entities/WalletModel.java
@@ -32,7 +32,7 @@ public class WalletModel {
     private double quantity;
 
     @Builder
-    public WalletModel(AccountModel account, CoinModel coin, float quantity) {
+    public WalletModel(AccountModel account, CoinModel coin, double quantity) {
         this.account = account;
         this.coin = coin;
         this.quantity = quantity;
